Remove the attachable option to mip_net-federated
This option allowed worker node containers to connect and see the
services connected on that network, on all the federation nodes.

By changing how Exareme connects to the LDSM, we do not need this
anymore, which improves security, and allows us to guarantee the only
way to access local data is through the Exareme worker service.